Move helper funcs to utils.go, xcodebuild.go and xcpretty.go files.
