+++ utils.go
@@ -1,0 +1,141 @@
+package main
+
+import (
+	"errors"
+	"fmt"
+	"os"
+	"path"
+	"path/filepath"
+	"regexp"
+
+	"github.com/bitrise-io/go-utils/colorstring"
+	"github.com/bitrise-io/go-utils/command"
+	"github.com/bitrise-io/go-utils/fileutil"
+	"github.com/bitrise-io/go-utils/log"
+	"github.com/bitrise-io/go-utils/pathutil"
+	"github.com/bitrise-io/go-utils/stringutil"
+	cmd "github.com/bitrise-steplib/steps-xcode-test/command"
+)
+
+func isStringFoundInOutput(searchStr, outputToSearchIn string) bool {
+	r, err := regexp.Compile("(?i)" + searchStr)
+	if err != nil {
+		log.Warnf("Failed to compile regexp: %s", err)
+		return false
+	}
+	return r.MatchString(outputToSearchIn)
+}
+
+func saveRawOutputToLogFile(rawXcodebuildOutput string, isRunSuccess, didLogToStdout bool) (string, error) {
+	tmpDir, err := pathutil.NormalizedOSTempDirPath("xcodebuild-output")
+	if err != nil {
+		return "", fmt.Errorf("failed to create temp dir, error: %s", err)
+	}
+	logFileName := "raw-xcodebuild-output.log"
+	logPth := filepath.Join(tmpDir, logFileName)
+	if err := fileutil.WriteStringToFile(logPth, rawXcodebuildOutput); err != nil {
+		return "", fmt.Errorf("failed to write xcodebuild output to file, error: %s", err)
+	}
+
+	if !isRunSuccess || !didLogToStdout {
+		deployDir := os.Getenv("BITRISE_DEPLOY_DIR")
+		if deployDir == "" {
+			return "", errors.New("no BITRISE_DEPLOY_DIR found")
+		}
+		deployPth := filepath.Join(deployDir, logFileName)
+
+		if err := command.CopyFile(logPth, deployPth); err != nil {
+			return "", fmt.Errorf("failed to copy xcodebuild output log file from (%s) to (%s), error: %s", logPth, deployPth, err)
+		}
+		logPth = deployPth
+	}
+
+	if err := cmd.ExportEnvironmentWithEnvman("BITRISE_XCODE_RAW_TEST_RESULT_TEXT_PATH", logPth); err != nil {
+		log.Warnf("Failed to export: BITRISE_XCODE_RAW_TEST_RESULT_TEXT_PATH, error: %s", err)
+	}
+	return logPth, nil
+}
+
+func saveAttachments(scheme, testSummariesPath, attachementDir string) error {
+	if exist, err := pathutil.IsDirExists(attachementDir); err != nil {
+		return err
+	} else if !exist {
+		return fmt.Errorf("no test attachments found at: %s", attachementDir)
+	}
+
+	if found, err := UpdateScreenshotNames(testSummariesPath, attachementDir); err != nil {
+		log.Warnf("Failed to update screenshot names, error: %s", err)
+	} else if !found {
+		return nil
+	}
+
+	// deploy zipped attachments
+	deployDir := os.Getenv("BITRISE_DEPLOY_DIR")
+	if deployDir == "" {
+		return errors.New("no BITRISE_DEPLOY_DIR found")
+	}
+
+	zipedTestsDerivedDataPath := filepath.Join(deployDir, fmt.Sprintf("%s-xc-test-Attachments.zip", scheme))
+	if err := cmd.Zip(filepath.Dir(attachementDir), filepath.Base(attachementDir), zipedTestsDerivedDataPath); err != nil {
+		return err
+	}
+
+	if err := cmd.ExportEnvironmentWithEnvman("BITRISE_XCODE_TEST_ATTACHMENTS_PATH", zipedTestsDerivedDataPath); err != nil {
+		log.Warnf("Failed to export: BITRISE_XCODE_TEST_ATTACHMENTS_PATH, error: %s", err)
+	}
+
+	log.Donef("The zipped attachments are available in: %s", zipedTestsDerivedDataPath)
+	return nil
+}
+
+func getSummariesAndAttachmentPath(testOutputDir string) (testSummariesPath string, attachmentDir string, err error) {
+	const testSummaryFileName = "TestSummaries.plist"
+	if exist, err := pathutil.IsDirExists(testOutputDir); err != nil {
+		return "", "", err
+	} else if !exist {
+		return "", "", fmt.Errorf("no test logs found at: %s", testOutputDir)
+	}
+
+	testSummariesPath = path.Join(testOutputDir, testSummaryFileName)
+	if exist, err := pathutil.IsPathExists(testSummariesPath); err != nil {
+		return "", "", err
+	} else if !exist {
+		return "", "", fmt.Errorf("no test summaries found at: %s", testSummariesPath)
+	}
+
+	var attachementDir string
+	{
+		attachementDir = filepath.Join(testOutputDir, "Attachments")
+		if exist, err := pathutil.IsDirExists(attachementDir); err != nil {
+			return "", "", err
+		} else if !exist {
+			return "", "", fmt.Errorf("no test attachments found at: %s", attachementDir)
+		}
+	}
+
+	log.Debugf("Test summaries path: %s", testSummariesPath)
+	log.Debugf("Attachment dir: %s", attachementDir)
+	return testSummariesPath, attachementDir, nil
+}
+
+func printLastLinesOfRawXcodebuildLog(rawXcodebuildOutput string, isRunSuccess bool) {
+	const lastLines = "\nLast lines of the build log:"
+	if !isRunSuccess {
+		log.Errorf(lastLines)
+	} else {
+		log.Infof(lastLines)
+	}
+
+	fmt.Println(stringutil.LastNLines(rawXcodebuildOutput, 20))
+
+	if !isRunSuccess {
+		log.Warnf("If you can't find the reason of the error in the log, please check the raw-xcodebuild-output.log.")
+	}
+
+	log.Infof(colorstring.Magenta(`
+The log file is stored in $BITRISE_DEPLOY_DIR, and its full path
+is available in the $BITRISE_XCODE_RAW_TEST_RESULT_TEXT_PATH environment variable.
+
+If you have the Deploy to Bitrise.io step (after this step),
+that will attach the file to your build as an artifact!`))
+}
